requirements.txt
################  Requirements for code rose  ######################

numpy
opencv-python
pillow

